+++ tests/testthat/test-Data-validity.R
@@ -217,12 +217,12 @@ test_that("v_data_mixture: error for nObsshare of wrong length", {
 # v_data_DA ----
 
 test_that("v_data_DA passes for valid object", {
-  object <- h_get_data_augmented()
+  object <- h_get_data_da()
   expect_true(v_data_DA(object))
 })
 
 test_that("v_data_DA: error for Tmax of wrong length, negative values", {
-  object <- h_get_data_augmented()
+  object <- h_get_data_da()
 
   # We assign Tmax of wrong length and negative values.
   object@Tmax <- c(-10, -20)
@@ -234,7 +234,7 @@ test_that("v_data_DA: error for Tmax of wrong length, negative values", {
 })
 
 test_that("v_data_DA: error for u of wrong length and values", {
-  object <- h_get_data_augmented()
+  object <- h_get_data_da()
 
   # We assign vector u of length different than object@nObs
   # with some negative values and some other, greater than Tmax.
@@ -247,7 +247,7 @@ test_that("v_data_DA: error for u of wrong length and values", {
 })
 
 test_that("v_data_DA: error for t0 of wrong length, negative values", {
-  object <- h_get_data_augmented()
+  object <- h_get_data_da()
 
   # We assign vector t0 of length different than object@nObs
   # and some negative, non-sorted values.